Update condition checks and clean up code in profile page
Improved the check for the existence of server data in user profile details to consider an empty array. Cleaned up and compacted the error handling within Promise blocks for fetching user data. This prevents unwanted behavior due to the loadState not being correctly updated. The date displaying format was also slightly adjusted to enhance readability.